vyapaars-backend/requirements.txt

fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
psycopg2-binary
python-dotenv
alembic
